Break up convolution and subsampling layers.

Former-commit-id: 17fe827ed219fa1de5e39c1cb0f13c0b73f0cf71

+++ deeplearning4j-core/src/main/java/org/deeplearning4j/plot/FilterRenderer.java
@@ -167,7 +167,7 @@ public class FilterRenderer {
 
             for (int col = 0; col < data.columns(); col++ ) {
 
-                double matrix_value = (double) data.getScalar( row, col ).getDouble(0);
+                double matrix_value = data.getScalar( row, col ).getDouble(0);
 
                 // at this point we need round values into bins
